refactor : todo 객체 생성
