+++ src/app.js
@@ -1,6 +1,6 @@
 import {$userCreateButton, $userList, $todoInput} from './todoDOM.js';
 import {loadUser, userState, newUser} from './components/user.js';
-import {addTodoItem} from './components/todoInput.js';
+import {todo} from './components/todo.js';
 
 
 export const app = () => {
@@ -9,6 +9,6 @@ export const app = () => {
 
   $userCreateButton.addEventListener('click', newUser.handler);
   $userList.addEventListener('click', userState.change);
-  $todoInput.addEventListener('keyup', addTodoItem);
+  $todoInput.addEventListener('keyup', todo.addItem);
 }
