+++ src/template.js
@@ -5,5 +5,16 @@ export const template = {
 
     userButtons (name, userId) {
         return `<button class="ripple" data-userId="${userId}">${name}</button>`
+    },
+
+    todo (contents) {
+        return `<li class=">
+            <div class="view">
+                <input class="toggle" type="checkbox"/>
+                <label class="label">${contents}</label>
+                <button class="destroy"></button>
+            </div>
+            <input class="edit" value="${contents}" />
+        </li>`
     }
 }
